fix(audit): close v0.2.46 re-audit H-01..M-04 and L-02..L-04 (v0.2.47)
Anchor-gated framer decode, resampler lookahead, signed framer cadence, circular FIR, Ed25519 manifest verify via libsodium, DDC Nyquist/clamp guards, README and bench honesty.

+++ src/dsp/P25Phase2Framer.cpp
@@ -176,7 +176,7 @@ void P25Phase2Framer::tryEmitSuperframe()
                 emitSuperframeFragment(sync1, sync2, 0, sync2Offset, true);
                 m_synchronized = false;
                 m_inSyncAllowance = 0;
-                m_superframeDibitsSinceEmit = static_cast<uint64_t>(sync2Offset);
+                m_superframeDibitsSinceEmit = sync2Offset;
                 return;
             }
 
@@ -193,7 +193,7 @@ void P25Phase2Framer::tryEmitSuperframe()
                 emitSuperframeFragment(sync1, sync2, sync1Offset, sync1Offset, false);
                 m_synchronized = false;
                 m_inSyncAllowance = 0;
-                m_superframeDibitsSinceEmit = static_cast<uint64_t>(sync1Offset);
+                m_superframeDibitsSinceEmit = sync1Offset;
                 return;
             }
 
@@ -205,7 +205,7 @@ void P25Phase2Framer::tryEmitSuperframe()
                     emitSuperframeFragment(sync1, sync2, sync1Offset, sync2Offset, true);
                     m_synchronized = false;
                     m_inSyncAllowance = 0;
-                    m_superframeDibitsSinceEmit = static_cast<uint64_t>(sync2Offset);
+                    m_superframeDibitsSinceEmit = sync2Offset;
                     return;
                 }
             }
